pyproject.toml
[tool.poetry]
name = "packages-test"
version = "0.1.0"
description = ""
authors = ["gbc-collib <collinstasiak@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
gitpython = "^3.1.43"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
boto3 = "^1.35.81"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
